requirements.txt
fastapi
uvicorn[standard]

# Langchain
langchain
langchain-openai
langchain-google-genai

# Supabase
supabase

# Utilities
python-dotenv
pydantic
